requirements.txt
numpy==2.2.5
pillow==11.2.1
imageio==2.37.0